Add validation and logging

+++ src/idea/idea.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, HttpException, HttpStatus } from '@nestjs/common';
 import { Repository } from 'typeorm';
 import { IdeaEntity } from './idea.entity';
 import { IdeaDto } from './idea.dto';
@@ -17,22 +17,37 @@ export class IdeaService {
     }
 
     async createIdea(data: IdeaDto){
+        
         const idea = await this.ideaRepository.create(data);
         await this.ideaRepository.save(idea);
         return idea;
     }
 
     async readIdea(id: string){
-        return await this.ideaRepository.findOne({ where: { id }})
+        const idea = await this.ideaRepository.findOne({ where: { id }})
+        if (!idea) throw new HttpException('Not Found', HttpStatus.NOT_FOUND);
+
+        return idea;
     }
  
     async updateIdea(id: string, data: Partial<IdeaDto>){
-        await this.ideaRepository.update({id}, data);
-        return await this.ideaRepository.findOne({id});
+        let idea = await this.ideaRepository.findOne({ where: { id }})
+        if (!idea) throw new HttpException('Not Found', HttpStatus.NOT_FOUND);
+        
+        const valor = await this.ideaRepository.update({id}, data);
+
+        // console.log('valor');
+        // console.log(valor);
+
+        idea = await this.ideaRepository.findOne({ where: { id }})
+        return idea;
     }
 
     async deleteIdea(id: string){
+        const idea = await this.ideaRepository.findOne({ where: { id }})
+        if (!idea) throw new HttpException('Not Found', HttpStatus.NOT_FOUND);
+        
         await this.ideaRepository.delete({id});
-        return {delete: true};
+        return idea;
     }
 }
